fix DOCX export page border - interoperabillity. export case #2
This is a fix for the 2nd case of export of a page's border.
The case is when a border's distance from text is larger than 31pt.
In such a case - LO used to write the value as-is,
which caused a problem in Word.
Now the fix checks the value, and if it is larger than 31pt -
it converts it to a distance from the page margin.
Based on the problem described here:
http://wiki.openoffice.org/wiki/Writer/MSInteroperability/PageBorder

Change-Id: I79f721adc71ac744eb332fbf3fea8070e41ddabc
Reviewed-on: https://gerrit.libreoffice.org/4959