Updated core Project: dictionaries  a32a341efdbb9e53a7b832262a41f6fb6454bedf
Downgraded the Dutch spellchecker to version 2.00 ( fdo#63033 fdo#65994 )

After some problems with the upgrade to version 2.10g on Linux and Windows downgrading is better.
I think downgrading is better because I had the old working files still saved and that was working.
Maybe in the furture I will try again to make the upgrade after some more testing.

Change-Id: I2e21a40f03b44a336355500bb617865ca2704125
Reviewed-on: https://gerrit.libreoffice.org/4987
Reviewed-by: Fridrich Strba <fridrich@documentfoundation.org>
Tested-by: Fridrich Strba <fridrich@documentfoundation.org>