let's own the packaged stuff by root:root again (fdo#67388)
We have started to use -fvisibility=hidden globally with GCC.
The hacky libgetuid.so does not work if the symbols are hidden.

In addition, the library was renamed from getuid.so to libgetuid.so
during gbuildisation; There was more location where it was not fixed;
The result was that the packages and files in the Linux download tarballs
were owned by some "ugly" user and group.

Change-Id: I22f62b861c373f2553dcccc85a84dddabaf1f418