Adapt UnoControlModel::read/write
...to 7786399d5fbe4f8507633212087a8133bda6f110 "fdo#67235 adapt form control
code to time nanosecond API change, step 3."  It is a bit unclear to me how
exactly this code is used, so to be safe, just read and write in the old format
(of using a single integer to represent a Date resp. Time) at least for now,
loosing nanosecond precision and the UTC flag.

Change-Id: Ib5148f750a420ad09366c79b68370ad0efd501f4