fdo#62442: Move the InfoPlist.strings files into the app from langpacks
The digital signature of an app bundle includes an integrity check for
the Resources subtree. This is the normal place for Mac apps to have
their read-only non-code "resources". In LibreOffice it is currently
not much used, though.

The signature thus breaks when a lang pack installer is run and plops
its InfoPlist.strings file into the "Resources" subtree. This file
contains translations of strings in the app's Info.plist file. For
LibreOffice, it contains translations only for some of the file type
names in Info.plist. (Why only some, I don't know.)

Files installed by a lang pack insaller into other locations in the
app bundle don't harm the signature.

Making the InfoPlist.strings files be distributed as part of the main
app bundle instead of in langpacks should keep the signature valid
even if a lang pack is installed. The InfoPlist.strings files are
small so the size of the main app should not grow significantly.

This required introduction of a new functionality in scp2: The
possibility to generate a list of several complete File or Directory
stanzas, one for each language for which translations are being
built.

(This is different from the existing functionality, used for files
that go into lang packs, where a stanza contains several Name or
HostName attributes (or whatever the term is) where the attribute
names are qualified with the language code in parens.)

Change-Id: I3afd9b08944fe1bccb5f0c881d740d260f589e39