Refactoring: remove 3 unneeded ivars
The default font size, font name and font traits were stored as ivars
of AquaA11yWrapper, but they are in fact only needed as temporary state
for the createAttributedStringForElement:inOrigRange: method of
AquaA11yTextAttributesWrapper. So remove these 3 ivars and instead
introduce a class that holds these 3 properties and make
createAttributedStringForElement:inOrigRange: use instance of this
class to hold the needed state instead.

I checked that the default font size, font name and font traits ivars
are really only used in AquaA11yTextAttributesWrapper at that one place.

Change-Id: Id2e45977c394db116f3fb0636136300c23e71f25
Reviewed-on: https://gerrit.libreoffice.org/5346
Reviewed-by: Tor Lillqvist <tml@iki.fi>
Tested-by: Tor Lillqvist <tml@iki.fi>