DOCX import: fixed horizontal position of floating tables
In Word, just like normal tables, floating tables should be positioned
in a way that the start of the cell text has the same horizontal
position as normal paragraph text.

To emulate this, first the table should be moved left by the table
border distance, then also by the border with / 2; as done for
non-floating tables already.

Change-Id: I581311fbb08009e6c1839106e8f615d078a4a705