Revert "Fixing i#119950 warn about other running terminal sessions"
This reverts commit b068cf122bfefa99199f9145bcc95f518eb3edda.

The original patch just looks for another instance running which
of course can be in the same terminal session. The assumptions
the author makes are wrong.

The bug the patch tries to solve is real and we'll probably
have to solve it in the future, but not this way.

Conflicts:

	desktop/source/app/officeipcthread.cxx

Change-Id: I1e645671aa09c6147163820c12d5efb3b3688bd1
Reviewed-on: https://gerrit.libreoffice.org/5415
Reviewed-by: Fridrich Strba <fridrich@documentfoundation.org>
Tested-by: Fridrich Strba <fridrich@documentfoundation.org>