Revert "Revert "Always disable anti-aliasing for drawing cell borders." fdo#60805"
This reverts commit 4c437f3f8b9e057c4448cc1328bdca0c47a714dd.

Unfortunately this re-introduces the previous problem of the cell borders
drawn too thick when anti-aliasing is enabled.  Also, even with this reverted
the problem will still persist if the user manually disables anti-aliasing
in the Options dialog. So, we *do* need to find a way to draw the
double-borders etc correctly without anti-aliasing.