DOCX import: handle w:leftFromText and w:rightFromText for floating tables
Left margin wasn't implemented, that's simple. Right margin of the table
was set to the cell margin in commit
3fa7ad80e923ac1c5352247e73cb59db3d15f53b, which turns out to be wrong:
it's true that the right margin should be >0, but not because of the
cell margin but because of the table margin. The new behavior matches
what the binary import always did.

Change-Id: Ifc24e4f086c49d5d575defdfca1d27e497fa03dc