rtl::compareAsciiIgnoreCase cannot be used here
...as its assert requires that both input characters are ASCII, which need not
be the case in these compareIgnoreAsciiCase functions.  (Even if they take one
literal argument that must be strictly ASCII, the other argument can be an
arbitrary Unicode string in the case of OUString or an arbitrary 8-bit string in
the case of OString).

The logically correct version of rtl::compareAsciiIgnoreCase would arguably be
one that requires its two arguments to be valid UTF-32 code units, but that
could not be used in these places either, as for OUString they operate on
individual UTF-16 code units.

rtl::compareAsciiIgnoreCase likely makes less sense after all than assumed in
1c883c2357c46f1db34bb8a58f8ac1d201cee403 "Introduce rtl::compareIgnoreCase and
deprecate rtl/character.hxx equivalents," which this commit partly reverts.

Change-Id: Ib2eed3a1896e83d9c66b0479a03f9ec51e1c4dc0