Fix borderline primitive merge
Two problems
1, Borderlines with different styles are merged
Steps to reproduce this bug:
-Add solid/dotted/dashed border to the first paragraph
-Add solid/dotted/dashed border (but different from the
first paragraph's border) to the second paragraph.
Result:
Second paragraph's left and right border will have the
same style as the first paragraph's border has. Top and bottom
border have the expected style.

2, The order of adding border lines to the BorderLines
container can change the correctness of merge.
The bug appaers when there is a vertical text
range with border, consist of more text portions.
For example: Add one border for the whole vertical text
range, but change the background of one part of it.

Change-Id: Iae6ecff31f24b232b3250c3f9d0d4ecd49fb18d4