Resolves: fdo#59048 crash on text-to-table undo
Delete the SwCellFrms attached to the layout of this SwTableNode early,
otherwise the SwCellFrms will dereference the SwTableNode after it is deleted
in _DeleteBox during the DelFrms of the DeleteSection in _DeleteBox when
invalidating following SwFrms.

SwUndoInsTbl::UndoImpl calls SwTableNode::DelFrms so presumably this is the
right way to go about it.

Change-Id: I6e088a30500d721bd18c57ced5c8ec331bb567eb
Reviewed-on: https://gerrit.libreoffice.org/5634
Reviewed-by: Michael Stahl <mstahl@redhat.com>
Tested-by: Michael Stahl <mstahl@redhat.com>