disable/disconnect style preview from styles dialog
So, the logic for firing the Preview start/end slots is disabled now,
I think the present dialog's behaviour could be very confusing, really
we need to have the preview done on the item the mouse is over (like
font preview) but I haven't had time to try and do that yet, also
imho it makes more sense to actually replace that styles dialog for
calc. I leave the code in place now so at least someone might be
able to enable and experiment with it.

Change-Id: I62adb9c217e9887d8df5ecea1e8ca071a570ab3e